@@ -1,5 +1,3 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
+require_relative 'application'
 
-# Initialize the rails application
-Dummy::Application.initialize!
+Rails.application.initialize!
